🐛 Fixed signbit problem on windows.

+++ src/hybridmap/AodScheduler.cpp
@@ -259,7 +259,8 @@ void AodScheduler::AodActivationHelper::reAssignOffsets(
             });
   int32_t offset = sign;
   for (auto* aodMove : aodMoves) {
-    if (std::signbit(aodMove->delta) == std::signbit(sign)) {
+    if (std::signbit(aodMove->delta) ==
+        std::signbit(static_cast<double>(sign))) {
       aodMove->offset = offset;
       offset += sign;
     }
@@ -400,7 +401,7 @@ uint32_t AodScheduler::AodActivationHelper::getMaxOffsetAtInit(
   uint32_t maxOffset = 0;
   for (const auto& aodMove : aodMoves) {
     auto offset = aodMove->offset;
-    if (std::signbit(offset) == std::signbit(sign)) {
+    if (std::signbit(offset) == std::signbit(static_cast<double>(sign))) {
       maxOffset = std::max(maxOffset, static_cast<uint32_t>(std::abs(offset)));
     }
   }
